Restructured Header component
